Server / Launcher
First pass of multi server done, need to check everything.
Launcher fix crash and fix correct progress bar value of total size

+++ Launcher/QDownloader/QDownloader.h
@@ -46,8 +46,8 @@ private:
     QFile output;
     QTime downloadTime;
 
-    int downloadedCount;
-    int totalCount;
+    float downloadedCount;
+    float totalCount;
 private slots:
     void startNextDownload();
